pyproject.toml
[project]
name = "openbot"
version = "0.1.0"
description = "Personal AI agent with multi-platform messaging, tool execution, and memory system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # LLM providers
    "anthropic>=0.52.0",
    "openai>=1.82.0",
    "dashscope>=1.22.0",
    # Data & config
    "pydantic>=2.11.0",
    "pydantic-settings>=2.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18.10",
    # HTTP client
    "httpx>=0.28.0",
    "qrcode[pil]>=8.2",
    "cryptography>=45.0.0",
    "lark-oapi>=1.5.3",
    # Logging
    "structlog>=25.1.0",
    # Database
    "aiosqlite>=0.21.0",
    "sqlite-vec>=0.1.6",
    # Telegram
    "python-telegram-bot>=22.6",
    # Scheduler
    "apscheduler>=3.11.0",
    # Web scraping
    "trafilatura>=2.0.0",
    # Search
    "tavily-python>=0.5.0",
    # Observability (optional OTel export)
    "opentelemetry-api>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.30.0",
]

[project.scripts]
openbot = "main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.11.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
